backend/api/api_fastapi/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "csearch-api-fastapi"
version = "0.1.0"
description = "FastAPI rewrite of the CSearch read API"
requires-python = ">=3.11"
dependencies = [
  "asyncpg>=0.29.0",
  "fastapi>=0.115.0",
  "openai>=1.0.0",
  "pydantic-settings>=2.3.0",
  "redis>=5.0.0",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

